==> Health2000-2011/6_Physical_measures/3_BMI.Rmd <==
---
title: "BMI"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `bmi`
* Variable label: `Body Mass Index calculated as weight (kg)/height(m)^2`
* Variable description: `the BMI variable should be created by dividing weight and height variables with the following equation= weight (kg)/height(m)^2`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process

## Finland

### Health2000-2011 - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `BMII_BMI`
* Label: `BMI1`
* Type of variable: `countinous`

```{r assign1}
opal.assign.table.tibble(o, 'Health20002011_w1_bmi','Health2000-2011.t00_athlos_2017_02_23',
                         variables=list('BMII_BMI'), missings = TRUE)
```

```{r local1}
Health20002011_w1_bmi <- opal.execute(o,'Health20002011_w1_bmi')
Continuous_summary(var = Health20002011_w1_bmi$BMII_BMI, missing_values = NA)
ggplot(Health20002011_w1_bmi, aes(BMII_BMI)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo1}
Health20002011_ds_w1 <- tibble(id=Health20002011_w1_bmi$id)
Health20002011_ds_w1$bmi <- car::recode(Health20002011_w1_bmi$BMII_BMI, "NA=999")
Health20002011_ds_w1$bmi <- labelled(Health20002011_ds_w1$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1}
Continuous_summary(var = Health20002011_ds_w1$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w1, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
It is going to check if weight and heigth values are correct by computing BMI with the formula: weight/height^2. 
```{r crosstabulation1}
opal.assign.table.tibble(o, 'Health20002011_w1_whb','Health2000-2011.t00_athlos_2017_02_23',
                         variables=list('KYS1_K17', 'BMII_PITUUS2', 'BMII_BMI'), missings = TRUE)
Health20002011_w1_whb <- opal.execute(o,'Health20002011_w1_whb')

Health20002011_w1_whb$bmi_c <- Health20002011_w1_whb$KYS1_K17/(Health20002011_w1_whb$BMII_PITUUS2)^2

Health20002011_w1_whb$proff <- Health20002011_w1_whb$BMII_BMI - Health20002011_w1_whb$bmi_c

summary(Health20002011_w1_whb$proff)

head(Health20002011_w1_whb[which(Health20002011_w1_whb$proff>2 | Health20002011_w1_whb$proff<(-2)),])

length(which(Health20002011_w1_whb$proff<0.8 & Health20002011_w1_whb$proff>-0.8))
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL1}
#val_labels(Health20002011_ds_w1) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w1', Health20002011_ds_w1)
#opal.symbol_import(o,'Health20002011_ds_w1', project='_Athlos_Harm_Dataset')
bmi <- Health20002011_ds_w1
save(bmi, file = "../RData/w1/bmi.RData")
rm(bmi)
```






### Health2000-2011 - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `BMII_BMI`
* Label: `BMI1`
* Type of variable: `countinous`

```{r assign2}
opal.assign.table.tibble(o, 'Health20002011_w2_bmi','Health2000-2011.t11_athlos_2017_02_27',
                         variables=list('BMII_BMI'), missings = TRUE)
```

```{r local2}
Health20002011_w2_bmi <- opal.execute(o,'Health20002011_w2_bmi')
Continuous_summary(var = Health20002011_w2_bmi$BMII_BMI, missing_values = NA)
ggplot(Health20002011_w2_bmi, aes(BMII_BMI)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo2}
Health20002011_ds_w2 <- tibble(id=Health20002011_w2_bmi$id)
Health20002011_ds_w2$bmi <- car::recode(Health20002011_w2_bmi$BMII_BMI, "NA=999")
Health20002011_ds_w2$bmi <- labelled(Health20002011_ds_w2$bmi, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2}
Continuous_summary(var = Health20002011_ds_w2$bmi, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w2, aes(bmi)) + geom_histogram(stat="count", fill="steelblue") + xlab("BMI") + ylab("Frequency") + xlim(10,100)
```

#### Validation
It is going to check if weight and heigth values are correct by computing BMI with the formula: weight/height^2. 
```{r crosstabulation2}
opal.assign.table.tibble(o, 'Health20002011_w2_whb','Health2000-2011.t11_athlos_2017_02_27',
                         variables=list('KYS1_K17', 'KYS1_K17B', 'BMII_BMI'), missings = TRUE)

Health20002011_w2_whb <- opal.execute(o,'Health20002011_w2_whb')

Health20002011_w2_whb$bmi_c <- Health20002011_w2_whb$KYS1_K17/(Health20002011_w2_whb$KYS1_K17B/100)^2

Health20002011_w2_whb$proff <- Health20002011_w2_whb$BMII_BMI - Health20002011_w2_whb$bmi_c

summary(Health20002011_w2_whb$proff)

head(Health20002011_w2_whb[which(Health20002011_w2_whb$proff>2 | Health20002011_w2_whb$proff<(-2)),])

length(which(Health20002011_w2_whb$proff<0.8 & Health20002011_w2_whb$proff>-0.8))
```
#### Quality estimation
This variable had to be derived.


#### Import to OPAL
```{r importOPAL2}
#val_labels(Health20002011_ds_w2) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w2', Health20002011_ds_w2)
#opal.symbol_import(o,'Health20002011_ds_w2', project='_Athlos_Harm_Dataset')
bmi <- Health20002011_ds_w2
save(bmi, file = "../RData/w2/bmi.RData")
rm(bmi)
```




```{r closeRsession, echo=FALSE}
opal.logout(o)
```
