==> Health2000-2011/6_Physical_measures/4_Obesity.Rmd <==
---
title: "Obesity"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ######################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ######################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `obesity`
* Variable label: `Presence of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2`
* Variable description: `the BMI variable should be created by dividing weight and hiegh variables by means of the following equation= weight (kg)/height(m)^2. Then, the Obesity variable should be generated according to the WHO cutoff point for BMI classification`
* Domain: `Physical measures`
* Value type: `categorical`
* Category coding:

**Code** | **Category Label**
-------- | ------------------
0        | not obese
1        | obese

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process


```{r lab, echo=FALSE}
ds_label <- "Precense of obesity based in Body Mass Index classification according to the WHO BMI classification (http://apps.who.int/bmi/index.jsp?introPage=intro_3.html): BMI value equal to or above 30Kg/m^2"
```


## Finland

### Wave 1

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `BMII_BMI`
* Label: `BMI1`
* Type of variable: `countinous`
* Missings
   + `NA` 

```{r read1, include=FALSE}
opal.assign.table.tibble(o, 'Health20002011_bmi','Health2000-2011.t00_athlos_2017_02_23',
                         variables=list('BMII_BMI'), missings = TRUE)
Health20002011_bmi <- opal.execute(o,'Health20002011_bmi')
```

```{r local11}
Continuous_summary(Health20002011_bmi[[2]])
ggplot(Health20002011_bmi, aes(x=Health20002011_bmi[[2]])) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm

To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo11}
Health20002011_ds_w1 <- tibble(id=Health20002011_bmi$id)
Health20002011_ds_w1$obesity <- car::recode(Health20002011_bmi[[2]], "lo:29.9999999=0; 30:hi=1; NA=999")
Health20002011_ds_w1$obesity <- labelled(Health20002011_ds_w1[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript11}
Categorical_summary(var = Health20002011_ds_w1[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w1, aes(x=factor(Health20002011_ds_w1[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab11}
BeforeH <- c(sum(na.omit(Health20002011_bmi[[2]])<30),sum(na.omit(Health20002011_bmi[[2]])>=30), sum(is.na(Health20002011_bmi[[2]])))
AfterH <- table(Health20002011_ds_w1[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","NA->999") 
kable(C)
```

#### Quality estimation
[Comments on the quality of the new harmonised variable. This section can be done in a later stage, once all variables are harmonised.]

```{r read, include=FALSE}
rm(Health20002011_bmi)
```


### Wave 2

#### Study-specific variable description

The study-specific variables elected to be harmonised are:

* Name: `BMII_BMI`
* Label: `BMI1`
* Type of variable: `countinous`
* Missings
   + `NA` 

```{r read2, include=FALSE}
opal.assign.table.tibble(o, 'Health20002011_bmi','Health2000-2011.t11_athlos_2017_02_27',
                         variables=list('BMII_BMI'), missings = TRUE)
Health20002011_bmi <- opal.execute(o,'Health20002011_bmi')
```

```{r local2}
Continuous_summary(Health20002011_bmi[[2]])
ggplot(Health20002011_bmi, aes(x=Health20002011_bmi[[2]])) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("BMI") + ylab("Frequency")
```

#### Harmonisation algorithm

To compute the harmonize variable from the study-specific variable it has to be recoded as follows:

* `BMI<30 into 0` 
* `BMI>=30 into 1`
* `NA into 999`

```{r harmo2}
Health20002011_ds_w2 <- tibble(id=Health20002011_bmi$id)
Health20002011_ds_w2$obesity <- car::recode(Health20002011_bmi[[2]], "lo:29.9999999=0; 30:hi=1; NA=999")
Health20002011_ds_w2$obesity <- labelled(Health20002011_ds_w2[[2]], labels = c("not obese"=0, "obese"=1,  "Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistical description of the new harmonised variable
```{r descript2}
Categorical_summary(var = Health20002011_ds_w2[[2]], missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w2, aes(x=factor(Health20002011_ds_w2[[2]]))) + geom_bar(stat="count", width=0.4, fill="steelblue") + xlab("Obesity") + ylab("Frequency")
```

#### Validation
```{r crosstab2}
BeforeH <- c(sum(na.omit(Health20002011_bmi[[2]])<30),sum(na.omit(Health20002011_bmi[[2]])>=30), sum(is.na(Health20002011_bmi[[2]])))
AfterH <- table(Health20002011_ds_w2[[2]], useNA = "ifany")
C <- rbind(BeforeH,AfterH)
colnames(C) <- c("<30->0",">=30->1","NA->999") 
kable(C)
```

#### Quality estimation
Once BMI has been calculated using weight (Kg) and Height (m^2),
Obesity was calculated by using the BMI cutoff>=30kg/m2 following the WHO 
classification. Finally, the resulting BMI variable was coded as follows: 
BMI>30kg/m2 =1 (obese); BMI<30kg/m2 =0 (not obese) to generate two categories.


```{r read22, include=FALSE}
rm(Health20002011_bmi)
```



<!--- ############################################# Import to OPAL ###########################################-->
```{r importOPAL, echo=FALSE}
obesity <- Health20002011_m_ds_w1   <- Health20002011_ds_w1
save(obesity, file = "../RData/w1/obesity.RData")
rm(obesity)

obesity <- Health20002011_m_ds_w2   <- Health20002011_ds_w2
save(obesity, file = "../RData/w2/obesity.RData")
rm(obesity)


l.hds <- list(Health20002011_ds_w1 = Health20002011_ds_w1 , Health20002011_ds_w2 = Health20002011_ds_w2)

m.hds <- list(Health20002011_m_ds_w1 = Health20002011_m_ds_w1 , Health20002011_m_ds_w2 = Health20002011_m_ds_w2)


for(i in 1:length(l.hds)) {
  label(l.hds[[i]][[2]]) <- label(m.hds[[i]][[2]]) <- ds_label # Text value assigned at the DS description, at the top.
    
  m.hds[[i]][[2]] <- car::recode(m.hds[[i]][[2]], "c(991, 992, 993, 994, 995, 996, 997,998, 999)=NA")
  m.hds[[i]][[2]] <- labelled(m.hds[[i]][[2]], labels = c("not obese"=0, "obese"=1, "Respond, died after this wave"=2, "Respond(proxy), died"=3, "No respond, alive"=4, "No respond, died this wave"=5, "No respond, died previous waves"=6, "No respond, dropped from sample"=7))

  nl <- names(l.hds[i])
  #opal.assign.data(o, nl , l.hds[[i]])
  #opal.symbol_import(o, nl , project='_Athlos_Harm_Dataset')  

  ml <- names(m.hds[i])
  #opal.assign.data(o, ml , m.hds[[i]]) 
  #opal.symbol_import(o, ml , project='_Athlos_Harm_Dataset')  

}
```



## Descriptive statistics of the harmonised variable accross populations and waves

Percentages of categories in each harmonised variable by cohort and wave:

```{r summ}
# Table HRS
t.hds <- frq(l.hds[[2]][2])[[1]][,c(1,2)] 
for (i in 1:length(l.hds)){
  t.hds[2+i] <- frq(l.hds[[i]][2])[[1]][,4] 
}
t.hds[13,] <- c("n", "sample size", 
             length(l.hds[[1]][[2]]), length(l.hds[[2]][[2]]))
names(t.hds)<-c("val", "label", "Wave1", "Wave2")
kable(t.hds)
```



```{r closeRsession, echo=FALSE}
opal.logout(o)
```
