==> Health2000-2011/6_Physical_measures/7_Grip_strength.Rmd <==
---
title: "Grip strength"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_laia.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `grip`
* Variable label: `Grip strength in kg`
* Variable description: `Grip strength`
* Domain: `Physical measures`
* Value type: `continuous`
* Comments: `According to a systematic review (Roberts et al, 2011), there are some inconsistencies in the use of maximum or mean grip strength as a summary measure which limit comparison of results between epidemiological studies. With multiple attempts, the maximum grip strength will be greater than the mean value. 
Even though for right handed people, there is a 10% stronger grip in the dominant hand, this is not true for left handed or ambidextrous people, whose grip strength are more or less equal in both hands (Roberts et al, 2011). 
In light of the above, the harmonised variable "grip" is proposed to be the highest score (to the nearest kg) of all hand grip measurements in each study, regardless of hand dominance. This is following a protocol suggested by Roberts et al (2011), based on the American Society of Hand Therapists (ASHT) protocol.`

* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

```{r lab, echo=FALSE}
ds_label <- "Grip strength"
```

# Data process

## Finland

### Wave 1 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `TOKY_4PURISTUS_MAX`| 
|-|-|
| **Label** | `Maximal grip strength, N`| 
| **Table name** | `t00_athlos_2017_02_23` | 
| **Description** |  

```{r assign1, echo=FALSE}
opal.assign.table.tibble(o, 'Health20002011_gs','Health2000-2011.t00_athlos_2017_02_23',
                         variables=list('TOKY_4PURISTUS_MAX'), missings = TRUE)
Health20002011_gs <- opal.execute(o,'Health20002011_gs')
```

```{r local1, echo=FALSE}
vari <- Health20002011_gs$TOKY_4PURISTUS_MAX
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(Health20002011_gs, aes(TOKY_4PURISTUS_MAX)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

**R script:**

```{r harmo1}
Health20002011_ds_w1 <- tibble(id=Health20002011_gs$id)
Health20002011_ds_w1$grip <- car::recode(Health20002011_gs$TOKY_4PURISTUS_MAX/10, "NA=999")
Health20002011_ds_w1$grip <- labelled(Health20002011_ds_w1$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1, echo=FALSE}
vari <- Health20002011_ds_w1$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(Health20002011_ds_w1, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation1, echo=FALSE}
```

```{r importOPAL1, echo=FALSE}
#val_labels(Health20002011_ds_w1) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w1', Health20002011_ds_w1)
#opal.symbol_import(o,'Health20002011_ds_w1', project='_Athlos_Harm_Dataset')
grip <- Health20002011_ds_w1
save(grip, file = "../RData/w1/grip.RData")
rm(grip)
```






### Wave 2 

#### Study-specific variable description

The study-specific variable elected to be harmonised is:

| **Name** | `PUR_MAX`| 
|-|-|
| **Label** | `The best of 1.-3. squeezes`| 
| **Table name** | `t11_athlos_2017_02_27` | 
| **Description** |  

```{r assign2, echo=FALSE}
opal.assign.table.tibble(o, 'Health20002011_gs','Health2000-2011.t11_athlos_2017_02_27',
                         variables=list('PUR_MAX'), missings = TRUE)
Health20002011_gs <- opal.execute(o,'Health20002011_gs')
```

```{r local2, echo=FALSE}
vari <- Health20002011_gs$PUR_MAX
kable(Continuous_summary(vari, missing_values = NA)[3], caption = ds_label) 
pander(Continuous_summary(vari, missing_values = NA)$summary, caption = "Summary")
ggplot(Health20002011_gs, aes(PUR_MAX)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the same than the study-specific variable, but the missing values have to be recoded as follows:

* `NA into 999`

**R script:**

```{r harmo2}
Health20002011_ds_w2 <- tibble(id=Health20002011_gs$id)
Health20002011_ds_w2$grip <- car::recode(Health20002011_gs$PUR_MAX, "NA=999")
Health20002011_ds_w2$grip <- labelled(Health20002011_ds_w2$grip, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2, echo=FALSE}
vari <- Health20002011_ds_w2$grip
kable(Continuous_summary(var = vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[3], caption = ds_label)
kable(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))[2], caption = "Values")
pander(Continuous_summary(vari, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))$summary, caption = "Summary")
ggplot(Health20002011_ds_w2, aes(grip)) + geom_histogram(stat="count", fill="steelblue") + xlab(ds_label) + ylab("Frequency") + xlim(0,100) 
```

#### Validation
```{r crosstabulation2, echo=FALSE}
```


```{r importOPAL2, echo=FALSE}
#val_labels(Health20002011_ds_w2) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w2', Health20002011_ds_w2)
#opal.symbol_import(o,'Health20002011_ds_w2', project='_Athlos_Harm_Dataset')
grip <- Health20002011_ds_w2
save(grip, file = "../RData/w2/grip.RData")
rm(grip)
```

### Quality estimation
At wave 1 the values are recoded from grams to kgs. 

```{r closeRsession, echo=FALSE}
opal.logout(o)
```
