==> Health2000-2011/6_Physical_measures/10_Systolic.Rmd <==
---
title: "Systolic"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `sbp`
* Variable label: `Systolic blood pressure of participant at each wave`
* Variable description: `Means of all sistolic blood pressure measures`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process

## Finland

### Health2000-2011 - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `MIT1_SYSTBP1`
* Label: `Systolic blood pressure 1st measurement`
* Type of variable: `countinous`

* Name: `MIT1_SYSTBP2`
* Label: `Systolic blood pressure 2nd measurement`
* Type of variable: `countinous`

```{r assign1}
opal.assign.table.tibble(o, 'Health20002011_w1_sbp','Health2000-2011.t00_athlos_2017_02_23',
                         variables=list('MIT1_SYSTBP1', 'MIT1_SYSTBP2'), missings = TRUE)
```

```{r local1}
Health20002011_w1_sbp <- opal.execute(o,'Health20002011_w1_sbp')

Continuous_summary(var = Health20002011_w1_sbp$MIT1_SYSTBP1, missing_values = NA)
ggplot(Health20002011_w1_sbp, aes(MIT1_SYSTBP1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = Health20002011_w1_sbp$MIT1_SYSTBP2, missing_values = NA)
ggplot(Health20002011_w1_sbp, aes(MIT1_SYSTBP2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo1}
Health20002011_w1_sbp$mean_sbp <- apply(X = Health20002011_w1_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

### NA only in the second one
idNA_only02 <- which(!is.na(Health20002011_w1_sbp$MIT1_SYSTBP1) & is.na(Health20002011_w1_sbp$MIT1_SYSTBP2))

Health20002011_w1_sbp$mean_sbp[idNA_only02] <- Health20002011_w1_sbp$MIT1_SYSTBP1[idNA_only02]

Health20002011_ds_w1 <- tibble(id=Health20002011_w1_sbp$id)
Health20002011_ds_w1$sbp <- car::recode(Health20002011_w1_sbp$mean_sbp, "NA=999")
Health20002011_ds_w1$sbp <- labelled(Health20002011_ds_w1$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1}
Continuous_summary(var = Health20002011_ds_w1$sbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w1, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation1}
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL1}
#val_labels(Health20002011_ds_w1) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w1', Health20002011_ds_w1)
#opal.symbol_import(o,'Health20002011_ds_w1', project='_Athlos_Harm_Dataset')
sbp <- Health20002011_ds_w1
save(sbp, file = "../RData/w1/sbp.RData")
rm(sbp)
```






### Health2000-2011 - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `MIT1_SYSTBP1`
* Label: `Systolic blood pressure 1st measurement`
* Type of variable: `countinous`

* Name: `MIT1_SYSTBP2`
* Label: `Systolic blood pressure 2nd measurement`
* Type of variable: `countinous`

```{r assign2}
opal.assign.table.tibble(o, 'Health20002011_w2_sbp','Health2000-2011.t11_athlos_2017_02_27',
                         variables=list('MIT1_SYSTBP1', 'MIT1_SYSTBP2'), missings = TRUE)
```

```{r local2}
Health20002011_w2_sbp <- opal.execute(o,'Health20002011_w2_sbp')

Continuous_summary(var = Health20002011_w2_sbp$MIT1_SYSTBP1, missing_values = NA)
ggplot(Health20002011_w2_sbp, aes(MIT1_SYSTBP1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = Health20002011_w2_sbp$MIT1_SYSTBP2, missing_values = NA)
ggplot(Health20002011_w2_sbp, aes(MIT1_SYSTBP2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency")
```

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo2}
Health20002011_w2_sbp$mean_sbp <- apply(X = Health20002011_w2_sbp[,c(2,3)], MARGIN = 1, FUN = mean)

### NA only in the second one
idNA_only02 <- which(!is.na(Health20002011_w2_sbp$MIT1_SYSTBP1) & is.na(Health20002011_w2_sbp$MIT1_SYSTBP2))

Health20002011_w2_sbp$mean_sbp[idNA_only02] <- Health20002011_w2_sbp$MIT1_SYSTBP1[idNA_only02]


Health20002011_ds_w2 <- tibble(id=Health20002011_w2_sbp$id)
Health20002011_ds_w2$sbp <- car::recode(Health20002011_w2_sbp$mean_sbp, "NA=999")
Health20002011_ds_w2$sbp <- labelled(Health20002011_ds_w2$sbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2}
Continuous_summary(var = Health20002011_ds_w2$sbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w2, aes(sbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Systolic Blood Pressure") + ylab("Frequency") + xlim(50,300) 
```

#### Validation
```{r crosstabulation2}
```

#### Quality estimation

Systolic bp was calculated as the mean for the morning, then the mean for the evening,
then the mean for both moments of the day, then the mean for the seven days in which 
this measure is taken to have a final value for this variable.

#### Import to OPAL
```{r importOPAL2}
#val_labels(Health20002011_ds_w2) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w2', Health20002011_ds_w2)
#opal.symbol_import(o,'Health20002011_ds_w2', project='_Athlos_Harm_Dataset')
sbp <- Health20002011_ds_w2
save(sbp, file = "../RData/w2/sbp.RData")
rm(sbp)
```




```{r closeRsession, echo=FALSE}
opal.logout(o)
```
