==> Health2000-2011/6_Physical_measures/11_Diastolic.Rmd <==
---
title: "Diastolic"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---

<!-- ########################################################## --> 

<!-- # Installation, loading packages -->
<!-- **R script:** -->

```{r setup, include=FALSE}
source("M:/WPs/WP1/Maelstrom/data_process/athlos-project.github.io/setup_iago.r")
```
<!-- ########################################################## --> 



# Description of DataSchema variable

The description of harmonised variable is the following:

* Short name: `dbp`
* Variable label: `Diastolic blood pressure of participantat at each wave`
* Variable description: `Means of all diastolic blood pressure measures`
* Domain: `Physical measures`
* Value type: `continuous`
* Category missings: 

**Code** | **Category Label**
----| ------------------
991 | CAPI/interviewer error
992 | Impute
993 | Disable to measure
994 | Not attempt/not done
995 | Does not answer
996 | Not applicable
997 | Refuse
998 | Do not know
999 | Missing

# Data process

## Finland

### Health2000-2011 - Wave 1 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `MIT1_DIASTBP1`
* Label: `Diastolic blood pressure 1st measurement`
* Type of variable: `countinous`

* Name: `MIT1_DIASTBP2`
* Label: `Systolic blood pressure 2nd measurement`
* Type of variable: `countinous`

```{r assign1}
opal.assign.table.tibble(o, 'Health20002011_w1_dbp','Health2000-2011.t00_athlos_2017_02_23',
                         variables=list('MIT1_DIASTBP1', 'MIT1_DIASTBP2'), missings = TRUE)
```

```{r local1}
Health20002011_w1_dbp <- opal.execute(o,'Health20002011_w1_dbp')

Continuous_summary(var = Health20002011_w1_dbp$MIT1_DIASTBP1, missing_values = 0)
ggplot(Health20002011_w1_dbp, aes(MIT1_DIASTBP1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = Health20002011_w1_dbp$MIT1_DIASTBP2, missing_values = 0)
ggplot(Health20002011_w1_dbp, aes(MIT1_DIASTBP2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

There are tree individuals with values 0 that are going to be considered as mitakes and will be recoded as 999 (missing). 

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo1}
Health20002011_w1_dbp$mean_dbp <- apply(X = Health20002011_w1_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

### NA only in the second one
idNA_only02 <- which(!is.na(Health20002011_w1_dbp$MIT1_DIASTBP1) & is.na(Health20002011_w1_dbp$MIT1_DIASTBP2))

Health20002011_w1_dbp$mean_dbp[idNA_only02] <- Health20002011_w1_dbp$MIT1_DIASTBP1[idNA_only02]

Health20002011_ds_w1 <- tibble(id=Health20002011_w1_dbp$id)
Health20002011_ds_w1$dbp <- car::recode(Health20002011_w1_dbp$mean_dbp, "NA=999; 0=999")
Health20002011_ds_w1$dbp <- labelled(Health20002011_ds_w1$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript1}
Continuous_summary(var = Health20002011_ds_w1$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w1, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
There are some unusual values on the variable MIT1_DIASTBP1: 
```{r crosstabulation1}
Health20002011_w1_dbp[which(Health20002011_w1_dbp$MIT1_DIASTBP1<30),]
```

#### Quality estimation
[Comments on the quality of the new harmonised variable.]

#### Import to OPAL
```{r importOPAL1}
#val_labels(Health20002011_ds_w1) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w1', Health20002011_ds_w1)
#opal.symbol_import(o,'Health20002011_ds_w1', project='_Athlos_Harm_Dataset')
dbp <- Health20002011_ds_w1
save(dbp, file = "../RData/w1/dbp.RData")
rm(dbp)
```






### Health2000-2011 - Wave 2 

#### Study-specific variables

The study-specific variable elected to be harmonised is:

* Name: `MIT1_DIASTBP1`
* Label: `Diastolic blood pressure 1st measurement`
* Type of variable: `countinous`

* Name: `MIT1_DIASTBP2`
* Label: `Systolic blood pressure 2nd measurement`
* Type of variable: `countinous`

```{r assign2}
opal.assign.table.tibble(o, 'Health20002011_w2_dbp','Health2000-2011.t11_athlos_2017_02_27',
                         variables=list('MIT1_DIASTBP1', 'MIT1_DIASTBP2'), missings = TRUE)
```

```{r local2}
Health20002011_w2_dbp <- opal.execute(o,'Health20002011_w2_dbp')

Continuous_summary(var = Health20002011_w2_dbp$MIT1_DIASTBP1, missing_values = 0)
ggplot(Health20002011_w2_dbp, aes(MIT1_DIASTBP1)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")

Continuous_summary(var = Health20002011_w2_dbp$MIT1_DIASTBP2, missing_values = 0)
ggplot(Health20002011_w2_dbp, aes(MIT1_DIASTBP2)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency")
```

There are tree individuals with values 0 that are going to be considered as mitakes and will be recoded as 999 (missing). 

#### Harmonisation algorithm
The harmonized variable is the mean value between the first and second measurement of the study-specific variables, but the missing values have to be recoded as follows:

* `NA into 999`

R code of the ALGORITHM:

```{r harmo2}
Health20002011_w2_dbp$mean_dbp <- apply(X = Health20002011_w2_dbp[,c(2,3)], MARGIN = 1, FUN = mean)

### NA only in the fisrt measurement
idNA_only01 <- which(is.na(Health20002011_w2_dbp$MIT1_DIASTBP1) & !is.na(Health20002011_w2_dbp$MIT1_DIASTBP2))

Health20002011_w2_dbp$mean_dbp[idNA_only01] <- Health20002011_w2_dbp$MIT1_DIASTBP2[idNA_only01]


### NA only in the second measurement
idNA_only02 <- which(!is.na(Health20002011_w2_dbp$MIT1_DIASTBP1) & is.na(Health20002011_w2_dbp$MIT1_DIASTBP2))

Health20002011_w2_dbp$mean_dbp[idNA_only02] <- Health20002011_w2_dbp$MIT1_DIASTBP1[idNA_only02]

### 0 only in the second measurement
id0_only02 <- which(Health20002011_w2_dbp$MIT1_DIASTBP1!=0 & Health20002011_w2_dbp$MIT1_DIASTBP2==0)

Health20002011_w2_dbp$mean_dbp[id0_only02] <- Health20002011_w2_dbp$MIT1_DIASTBP1[id0_only02]


Health20002011_ds_w2 <- tibble(id=Health20002011_w2_dbp$id)
Health20002011_ds_w2$dbp <- car::recode(Health20002011_w2_dbp$mean_dbp, "NA=999; 0=999")
Health20002011_ds_w2$dbp <- labelled(Health20002011_ds_w2$dbp, labels = c("Missing"=999, "Do not know"=998, "Refuse"=997, "Not applicable"=996, "Does not answer"=995,  "Not attempt not done"=994, "Disable to measure"=993, "Impute"=992, "CAPI interviewer error"=991))
```

#### Statistics of the new harmonised variable
```{r descript2}
Continuous_summary(var = Health20002011_ds_w2$dbp, missing_values = c(991, 992, 993, 994, 995, 996, 997, 998, 999))
ggplot(Health20002011_ds_w2, aes(dbp)) + geom_histogram(stat="count", fill="steelblue") + xlab("Diastolic Blood Pressure") + ylab("Frequency") + xlim(20,160) 
```

#### Validation
```{r crosstabulation2}
```

#### Quality estimation

The calculation of diastolic bp was reached by calculating the mean for the morning, 
then the mean for the evening, then the mean for both moments of the day, 
then the mean for the seven days in which this measure is taken to have a final value 
for this variable.

#### Import to OPAL
```{r importOPAL2}
#val_labels(Health20002011_ds_w2) <- NULL
#opal.assign.data(o, 'Health20002011_ds_w2', Health20002011_ds_w2)
#opal.symbol_import(o,'Health20002011_ds_w2', project='_Athlos_Harm_Dataset')
dbp <- Health20002011_ds_w2
save(dbp, file = "../RData/w2/dbp.RData")
rm(dbp)
```




```{r closeRsession, echo=FALSE}
opal.logout(o)
```
