==> Health2000-2011/6_Physical_measures/13_heart_rate.Rmd <==
---
title: "Heart Rate"
output:
  html_document:
    highlight: tango
    number_sections: yes
    theme: united
    toc: yes
    toc_depth: 6
    toc_float:
      collapsed: yes
      smooth_scroll: yes
---
  
<!-- ########################################################## --> 
<!-- # Installation, loading packages -->
```{r setup, include=FALSE}
source("../../setup.r")
```
<!-- ########################################################## --> 
<!-- # Loading DS description -->
```{r main, child='../../_DS_Rmd/heart_rateDS.Rmd'}
```
```{r global, echo=F}
source('../../_DS_Rmd/heart_rateDS.R')
```

# Data process


## Finland

### Wave 1

#### Study-specific variable description

| **Name** |`MIT1_HRATE`|
|-|-|
| **Label** |`Pulse rate per minute`|
| **Table name** |`t00_athlos_2017_02_23`|
| **Units** |`PULSE/min`|
| **Description** |  |




```{r assign w1, echo=F}

opal.assign.table.tibble(o, 'Health20002011_w1','Health2000-2011.t00_athlos_2017_02_23', variables=list('MIT1_HRATE'), missings = TRUE)
Health20002011_w1 <- opal.execute(o,'Health20002011_w1')

```

```{r local w1, echo=F}

vbl <-Health20002011_w1$MIT1_HRATE
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = "MIT1_HRATE") 
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Health20002011_w1, aes(MIT1_HRATE)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variable, it has to be recoded as follows:

* `NA into 999`

```{r harmo w1}

Health20002011_ds_w1 <- tibble(id=Health20002011_w1$id)
Health20002011_ds_w1$heart_rate <- car::recode(Health20002011_w1$MIT1_HRATE, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript w1, echo=F}

vbl <- Health20002011_ds_w1$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(Health20002011_ds_w1, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation w1, echo=F}
```





### Wave 2 

#### Study-specific variable description

| **Name** |`MIT1_HRATE`|
|-|-|
| **Label** |`Pulse rate per minute`|
| **Table name** |`t11_athlos_2017_02_27`|
| **Units** |`PULSE/min`|
| **Description** |  |




```{r assign w2, echo=F}

opal.assign.table.tibble(o, 'Health20002011_w2','Health2000-2011.t11_athlos_2017_02_27', variables=list('MIT1_HRATE'), missings = TRUE)
Health20002011_w2 <- opal.execute(o,'Health20002011_w2')

```

```{r local w2, echo=F}

vbl <-Health20002011_w2$MIT1_HRATE
kable(Continuous_summary(vbl, missing_values = NA)[3], caption = "MIT1_HRATE") 
pander(Continuous_summary(vbl, missing_values = NA)$summary, caption = "Summary")
ggplot(Health20002011_w2, aes(MIT1_HRATE)) + geom_histogram(stat="count", fill="steelblue") + xlab(label(vbl)) + ylab("Frequency")

```


#### Harmonisation algorithm

To compute `heart_rate` from the study-specific variable, it has to be recoded as follows:

* `NA into 999`

```{r harmo w2}

Health20002011_ds_w2 <- tibble(id=Health20002011_w2$id)
Health20002011_ds_w2$heart_rate <- car::recode(Health20002011_w2$MIT1_HRATE, "NA=999")

```


#### Statistics of the new harmonised variable
```{r descript w2, echo=F}

vbl <- Health20002011_ds_w2$heart_rate
kable(Continuous_summary(var = vbl, missing_values = miss_values_vector)[3], caption = ds_label)
kable(Continuous_summary(vbl, missing_values = miss_values_vector)[2], caption = "Values")
pander(Continuous_summary(vbl, missing_values = miss_values_vector)$summary, caption = "Summary")
ggplot(Health20002011_ds_w2, aes(heart_rate)) + geom_histogram(stat="count", fill="steelblue", data = .%>% filter(!vbl %in% miss_values_vector)) + xlab(ds_label) + ylab("Frequency")

```


#### Validation
```{r crosstabulation w2, echo=F}
```



## Summary of descriptive statistics of the harmonised variable accross populations and waves

```{r helpfunctions, echo=F}

labelling <- function(l.hds,m.hds){
  
  # Labelling of the tibbles with continuous data and creating new tibbles with all missings recodified as NA

  for(name in names(l.hds)) {
    # In the aux_object we copy the old tibble to recodify all missing values as NA.
    aux_object <- l.hds[[name]]
    # Labelling of variables
    label(l.hds[[name]][[2]]) <- label(aux_object[[2]]) <- ds_label
    # Labelling of categories (for continues variables, only missing values)
    l.hds[[name]][[2]] <- labelled(l.hds[[name]][[2]], labels = cont_label)
    aux_object[[2]] <- car::recode(aux_object[[2]], "miss_values_vector=NA")
    # Saving the recodified tibble in list m.hds
    m.hds[[name]] <- aux_object
    rm(aux_object)
  }
  return(list(l.hds,m.hds))

}

# Creation of summary tables for continuous data

summaries <- function(l.hds,m.hds,lnames){

  # Creation of column with summary table categories
  t.summ <- summary(m.hds[[1]][2])[1:6]
  # Adding of missing/no-missing values categories
  t.hds <- c(substr(t.summ,1,regexpr(":", t.summ, fixed=T)-1),labels(Continuous_summary(l.hds[[1]][[2]],missing_values = miss_values_vector)$values_table)[[2]])
  # For each wave/population in l.hds, add the correponding values
  for (i in seq_along(l.hds)){
    # First, summary values
    t.summ <- summary(m.hds[[i]][2])[1:6]
    # Next, missing/no-missing values
    t.hds <- cbind(t.hds,c(as.numeric(substr(t.summ,regexpr(":", t.summ, fixed=T)+1,nchar(t.summ))),as.numeric(Continuous_summary(l.hds[[i]][[2]],missing_values = miss_values_vector)$values_table[1,])))
  }
  # Add sample size for each wave/population
  t.hds <- rbind(t.hds,c("sample size", sapply(l.hds,function(wave) length(wave[[1]]))))
  # Add wave/population names
  dimnames(t.hds)[[2]] <- c(dimnames(summary(m.hds[[1]][2]))[[2]],lnames)
  return(t.hds)
  
}

# Save data tables

savingRD <- function(l.hds,vbl_name){
  
  for(index in seq_along(l.hds)){
    assign(vbl_name,l.hds[[index]])
    save(vbl_name,list = vbl_name, file = paste0(datafolder,names(l.hds)[index],"/",vbl_name,".RData"))
  }

}

```





```{r summ, echo=F}

# All study waves and populations with abbreviated and descriptive names


health2000.cw <- list(w1 = c("w1","W1"), w2 = c("w2","W2"))

# Consider only harmonised waves
l.hds <- lapply(health2000.cw, function(wname) if(exists(paste0("Health20002011_ds_",wname[1]))){wname = list(get(paste0("Health20002011_ds_",wname[1])),wname[2])})
nullw <- which(sapply(l.hds, is.null))

if(length(nullw)!=0){
  l.hds <- l.hds[-nullw]
}

m.hds <- list()

lmlist <- list(lapply(l.hds,function(x)x[[1]]),m.hds)

if(length(l.hds)>0){
  # Labelling tibbles
  lmlist <- labelling(lmlist[[1]],lmlist[[2]])
  # Printing summaries
  print(knitr::kable(summaries(lmlist[[1]],lmlist[[2]],lapply(l.hds,function(x)x[[2]])),caption=ds_label))
  # Saving tibbles
  savingRD(lmlist[[1]],vbl_name=hd_vbl)
}



```





# Quality estimation




<!-- ########################################################## -->
<!--- # Close OPAL R Session -->
```{r closeRsession, echo=FALSE}
opal.logout(o)
```
